requirements.txt
torch==2.4.1
torchvision
torchaudio
lightning==2.3.3
librosa
soundfile
tqdm
gdown
tensorboard
transformers
pandas
jiwer
openai-whisper